app/requirements.txt
flask==2.3.2
gunicorn==20.1.0
mysql-connector==2.2.9
